Doctest with MyCustomModule
